+++ tests/test_end_scenario_sizing.py
@@ -5,9 +5,9 @@ import mesido._darcy_weisbach as darcy_weisbach
 from mesido.esdl.esdl_parser import ESDLFileParser
 from mesido.esdl.profile_parser import ProfileReaderFromFile
 from mesido.workflows import (
-    EndScenarioSizingDiscountedHIGHS,
-    EndScenarioSizingHIGHS,
-    EndScenarioSizingStagedHIGHS,
+    EndScenarioSizing,
+    EndScenarioSizingDiscounted,
+    EndScenarioSizingStaged,
     run_end_scenario_sizing,
 )
 from mesido.workflows.grow_workflow import EndScenarioSizingHeadLossStaged
@@ -29,7 +29,7 @@ class TestEndScenarioSizing(TestCase):
         # This is an optimization done over a full year with timesteps of 5 days and hour timesteps
         # for the peak day
         cls.solution = run_optimization_problem(
-            EndScenarioSizingHIGHS,
+            EndScenarioSizing,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
             esdl_parser=ESDLFileParser,
@@ -133,7 +133,7 @@ class TestEndScenarioSizing(TestCase):
         solution_unstaged = self.solution
 
         solution_unstaged_2 = run_end_scenario_sizing(
-            EndScenarioSizingHIGHS,
+            EndScenarioSizing,
             staged_pipe_optimization=False,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
@@ -143,7 +143,7 @@ class TestEndScenarioSizing(TestCase):
         )
 
         solution_staged = run_end_scenario_sizing(
-            EndScenarioSizingStagedHIGHS,
+            EndScenarioSizingStaged,
             base_folder=base_folder,
             esdl_file_name="test_case_small_network_with_ates_with_buffer_all_optional.esdl",
             esdl_parser=ESDLFileParser,
@@ -240,7 +240,7 @@ class TestEndScenarioSizing(TestCase):
 
         base_folder = Path(run_ates.__file__).resolve().parent.parent
 
-        class TestEndScenarioSizingDiscountedHIGHS(EndScenarioSizingDiscountedHIGHS):
+        class TestEndScenarioSizingDiscountedHIGHS(EndScenarioSizingDiscounted):
             def solver_options(self):
                 options = super().solver_options()
                 options["solver"] = "highs"